ALFREDAPI-427 abstract out the type checking into its own service
